+++ forui/lib/widgets/select.dart
@@ -1,0 +1,4 @@
+/// Displays a list of options for the user to pick from.
+///
+/// TODO: This is a placeholder for the select widget.
+library forui.widgets.select;