+++ forui/lib/widgets/select_group.dart
@@ -6,5 +6,4 @@
 library forui.widgets.select_group;
 
 export '../src/widgets/select_group/select_group.dart';
-export '../src/widgets/select_group/select_group_controller.dart' hide DelegateSelectGroupController;
 export '../src/widgets/select_group/select_group_item.dart' hide FSelectGroupItemData;