+++ forui_hooks/lib/forui_hooks.dart
@@ -1,11 +1,14 @@
 export '/src/accordion_controller_hook.dart';
 export '/src/calendar_controller_hook.dart';
 export '/src/date_field_controller_hook.dart';
+export '/src/multi_value_notifier_hook.dart';
 export '/src/pagination_controller_hook.dart';
 export '/src/picker_controller_hook.dart';
 export '/src/popover_controller_hook.dart';
 export '/src/resizable_controller_hook.dart';
 export '/src/select_group_controller_hook.dart';
+export '/src/select_menu_tile_controller_hook.dart';
+export '/src/select_tile_group_controller_hook.dart';
 export '/src/slider_controller_hook.dart';
 export '/src/tab_controller_hook.dart';
 export '/src/time_field_controller_hook.dart';